Truncate author list
Technically, I'm the maintainer and lead developer, not the author of Minechem v5...but I have been the "point man" for about the past year, and linking to the full list of contributors seems like the best plan at this point (kinda tired of people thinking pixlepix is still writing it). 

Regardless of technicalities, I should be the first point of contact for anything related to Minechem...hence the edit.